Reorganise VFS mounting into TS module `mount.ts`
